saved work

+++ src/main/java/com/hcl/retailbanking/service/impl/ManagePayeeServiceImpl.java
@@ -6,35 +6,39 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.hcl.retailbanking.entities.ManagePayee;
-import com.hcl.retailbanking.pojos.CustomerCreation;
+
 
 import com.hcl.retailbanking.repositories.ManagePayeeRepository;
+import com.hcl.retailbanking.service.MailService;
 import com.hcl.retailbanking.service.ManagePayeeService;
 
 @Service
 public class ManagePayeeServiceImpl implements ManagePayeeService {
 	@Autowired
-	private ManagePayeeRepository managePayeeRepository;
-
-	@Override
-	public List<ManagePayee> getPayee(Long id) {
-		return managePayeeRepository.findAllById(id);
-	}
-
-	@Autowired
 	ManagePayeeRepository managePayeeRepo;
+	@Autowired
+	MailService mailService;
 
 	@Override
-	public void deletePayee(long payeeId) {
-		managePayeeRepo.deleteById(payeeId);
+	public String deletePayee(long payeeId) {
+		boolean payee = managePayeeRepo.existsById(payeeId);
+		if (payee) {
+			managePayeeRepo.deleteById(payeeId);
+			return "Payee deleted successfully";
+		} else
+			return "Payee doesn't exist";
 	}
 
 	@Override
 	public ManagePayee doPayeeValidation(ManagePayee payee) {
 		 ManagePayee payee1 = null;
 		if (payee.getPayee_id() != null) {
-			/*payee1 = managePayeeRepo.findById(payee.getId()).get();
-			if (payee1 == null) {*/
+			String generateOTP = MailService.generateOTP();
+			payee.setOtp(Integer.valueOf(generateOTP));
+			/*
+			 * payee1 = managePayeeRepo.findById(payee.getId()).get(); if (payee1 == null) {
+			 */
+			mailService.sendOTPMail(payee.getCustomer_email_id(), generateOTP);
 			payee1 = addPayee(payee);
 			// }
 		}
@@ -47,4 +51,13 @@ public class ManagePayeeServiceImpl implements ManagePayeeService {
 		return managePayeeRepo.save(payeeId);
 	}
 
+	@Override
+	public String verifyPayee(int otp, Long payeeId) {
+		ManagePayee verifiedPayee = managePayeeRepo.findById(payeeId).get();
+		if (verifiedPayee.getOtp() == otp) {
+			return "Payee verified successfully!";
+		} else {
+			return "Payee verification failed!";
+		}
+	}
 }
